+++ src/main/java/com/teamEWSN/gitdeun/Application/service/ApplicationService.java
@@ -5,9 +5,9 @@ import com.teamEWSN.gitdeun.Application.entity.Application;
 import com.teamEWSN.gitdeun.Application.entity.ApplicationStatus;
 import com.teamEWSN.gitdeun.Application.mapper.ApplicationMapper;
 import com.teamEWSN.gitdeun.Application.repository.ApplicationRepository;
-import com.teamEWSN.gitdeun.Recruitment.entity.Recruitment;
-import com.teamEWSN.gitdeun.Recruitment.entity.RecruitmentStatus;
-import com.teamEWSN.gitdeun.Recruitment.repository.RecruitmentRepository;
+import com.teamEWSN.gitdeun.recruitment.entity.Recruitment;
+import com.teamEWSN.gitdeun.recruitment.entity.RecruitmentStatus;
+import com.teamEWSN.gitdeun.recruitment.repository.RecruitmentRepository;
 import com.teamEWSN.gitdeun.common.exception.ErrorCode;
 import com.teamEWSN.gitdeun.common.exception.GlobalException;
 import com.teamEWSN.gitdeun.notification.dto.NotificationCreateDto;
