+++ src/main/java/com/teamEWSN/gitdeun/recruitment/entity/RecruitmentImage.java
@@ -1,4 +1,4 @@
-package com.teamEWSN.gitdeun.Recruitment.entity;
+package com.teamEWSN.gitdeun.recruitment.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
