+++ src/main/java/com/teamEWSN/gitdeun/recruitment/entity/RecruitmentRequiredSkill.java
@@ -1,4 +1,4 @@
-package com.teamEWSN.gitdeun.Recruitment.entity;
+package com.teamEWSN.gitdeun.recruitment.entity;
 
 import com.teamEWSN.gitdeun.userskill.entity.DeveloperSkill;
 import jakarta.persistence.*;
