+++ src/main/java/com/teamEWSN/gitdeun/recruitment/entity/SkillCategory.java
@@ -1,4 +1,4 @@
-package com.teamEWSN.gitdeun.Recruitment.entity;
+package com.teamEWSN.gitdeun.recruitment.entity;
 
 public enum SkillCategory {
     LANGUAGE,
