+++ src/main/java/com/teamEWSN/gitdeun/recruitment/repository/RecruitmentCustomRepository.java
@@ -1,8 +1,8 @@
-package com.teamEWSN.gitdeun.Recruitment.repository;
+package com.teamEWSN.gitdeun.recruitment.repository;
 
-import com.teamEWSN.gitdeun.Recruitment.entity.Recruitment;
-import com.teamEWSN.gitdeun.Recruitment.entity.RecruitmentField;
-import com.teamEWSN.gitdeun.Recruitment.entity.RecruitmentStatus;
+import com.teamEWSN.gitdeun.recruitment.entity.Recruitment;
+import com.teamEWSN.gitdeun.recruitment.entity.RecruitmentField;
+import com.teamEWSN.gitdeun.recruitment.entity.RecruitmentStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
