+++ src/main/java/com/teamEWSN/gitdeun/recruitment/repository/RecruitmentRequiredSkillRepository.java
@@ -1,6 +1,6 @@
-package com.teamEWSN.gitdeun.Recruitment.repository;
+package com.teamEWSN.gitdeun.recruitment.repository;
 
-import com.teamEWSN.gitdeun.Recruitment.entity.RecruitmentRequiredSkill;
+import com.teamEWSN.gitdeun.recruitment.entity.RecruitmentRequiredSkill;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
